// src/model/Node.ts
export interface Node {
    avatar_large: string
    name: string
    avatar_normal: string
    title: string
    url: string
    topics: number
    footer: string
    header: string
    title_alternative: string
    avatar_mini: string
    stars: number
    aliases: string[]
    root: boolean
    id: number
    parent_node_name: string
}